requirements.txt
asgiref==3.2.7
Django==3.0.5
image==1.5.30
Jinja2==2.11.2
MarkupSafe==1.1.1
numpy==1.18.2
opencv-python==4.2.0.34
pandas==1.0.3
Pillow==7.1.1
pytesseract==0.3.3
python-dateutil==2.8.1
pytz==2019.3
six==1.14.0
sqlparse==0.3.1
